# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VDM reference services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/connections/{connectionId}/devices/{deviceId}/latest": {
      "get": {
        "operationId": "latest",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local service state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Latest"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{connectionId}/devices/{deviceId}/alarms/local": {
      "get": {
        "operationId": "localAlarms",
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local service state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocalAlarm"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required when VDM_API_TOKEN is set; mandatory for remote bind"
      }
    },
    "schemas": {
...
      "Latest": {
        "type": "object",
        "required": [
          "telemetry",
          "attributes",
          "receivedCount"
        ],
        "properties": {
          "telemetry": {
            "type": "object",
            "nullable": true
          },
          "attributes": {
            "type": "object",
            "nullable": true
          },
          "receivedCount": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "LocalAlarm": {
        "type": "object",
        "required": [
          "alarmId",
          "eventId",
          "ts",
          "transition",
          "needsReconcile"
        ],
        "properties": {
          "alarmId": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          },
          "ts": {
            "description": "Device UTC seconds, number for JSON or decimal string for ProtoJSON",
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ]
          },
          "transition": {
            "type": "string"
          },
          "needsReconcile": {
            "type": "boolean"
          },
          "currentActive": {
            "type": "boolean"
          },
          "currentState": {
            "type": "object",
            "nullable": true
          },
          "reconciledAt": {
            "type": "number",
            "description": "Platform UTC seconds for state query; never substitutes device event time"
          }
        }
      }
    }
  }
}